Propagate KSJ ID
