+++ rust/src/lib.rs
@@ -7,9 +7,6 @@ use crate::{
     zip_reader::ZippedShapefileReader,
 };
 
-pub use crate::error::Ksj2GpError;
-pub use encoding::{decode_cp437cp932_to_utf8, encode_utf8_to_cp437cp932};
-
 mod builder;
 mod crs;
 mod encoding;
@@ -18,6 +15,10 @@ mod transform_coord;
 mod translate;
 mod writer;
 mod zip_reader;
+
+pub use crate::error::Ksj2GpError;
+pub use encoding::{decode_cp437cp932_to_utf8, encode_utf8_to_cp437cp932};
+pub use translate::{TranslateOptions, extract_ksj_id};
 
 pub fn list_shp_files<R: Read + Seek>(reader: R) -> Result<Vec<String>, Ksj2GpError> {
     match zip::ZipArchive::new(reader) {
@@ -41,6 +42,9 @@ pub fn convert_shp_inner<RW: Read + Seek + Write, R: Read + Seek, W: Write + Sen
     shx: RW,
     mut out: W,
     output_format: &str,
+    // Since `zip` is a file handle, it doesn't contain the filename. So, it
+    // needs to be extracted outside of this function.
+    translate_options: &TranslateOptions,
 ) -> Result<(), Ksj2GpError> {
     let mut zip = match zip::ZipArchive::new(zip) {
         Ok(zip) => ZippedShapefileReader::new(zip, target_shp),
@@ -62,26 +66,11 @@ pub fn convert_shp_inner<RW: Read + Seek + Write, R: Read + Seek, W: Write + Sen
 
     let mut reader = Reader::new(shapefile_reader, dbase_reader);
 
-    // TODO: pass this option from outside
-    let translate_colnames = true;
-    let translate_contents = true;
-
     match output_format {
-        "GeoParquet" => write_geoparquet(
-            &mut reader,
-            &mut out,
-            &dbf_fields,
-            &wkt,
-            translate_colnames,
-            translate_contents,
-        ),
-        "GeoJson" => write_geojson(
-            &mut reader,
-            &mut out,
-            &dbf_fields,
-            translate_colnames,
-            translate_contents,
-        ),
+        "GeoParquet" => {
+            write_geoparquet(&mut reader, &mut out, &dbf_fields, &wkt, translate_options)
+        }
+        "GeoJson" => write_geojson(&mut reader, &mut out, &dbf_fields, translate_options),
         _ => Err(format!("Unsupported format: {output_format}").into()),
     }
 }
