+++ rust/src/writer/geoparquet_writer.rs
@@ -6,7 +6,7 @@ use parquet::arrow::ArrowWriter;
 
 use crate::{
     builder::construct_schema, crs::wild_guess_from_esri_wkt_to_projjson, error::Ksj2GpError,
-    writer::get_fields_except_geometry,
+    translate::TranslateOptions, writer::get_fields_except_geometry,
 };
 
 // Number of rows to process at once
@@ -17,8 +17,7 @@ pub(crate) fn write_geoparquet<T: Read + Seek, D: Read + Seek, W: Write + Send>(
     writer: &mut W,
     dbf_fields: &[dbase::FieldInfo],
     wkt: &Option<String>,
-    use_readable_colnames: bool,
-    use_readable_contents: bool,
+    translate_options: &TranslateOptions,
 ) -> Result<(), Ksj2GpError> {
     let projjson = match wkt {
         Some(wkt) => wild_guess_from_esri_wkt_to_projjson(wkt)?,
@@ -27,7 +26,7 @@ pub(crate) fn write_geoparquet<T: Read + Seek, D: Read + Seek, W: Write + Send>(
     };
     let crs = geoarrow_schema::Crs::from_projjson(projjson);
 
-    let fields_info = construct_schema(dbf_fields, crs, use_readable_colnames)?;
+    let fields_info = construct_schema(dbf_fields, crs, translate_options)?;
     let schema_ref = fields_info.schema_ref.clone();
 
     // Since shapefile::Record is a HashMap, the iterator of it doesn't maintain
